+++ core/src/test/java/ru/kwanza/jeda/core/queue/TestTransactionalMemoryQueueWithDSTrx.java
@@ -335,7 +335,7 @@ public class TestTransactionalMemoryQueueWithDSTrx extends TestCase {
 
     public void testMaxSize() {
         TransactionalMemoryQueue memoryQueue = new TransactionalMemoryQueue(manager);
-        assertEquals("MaxSize wrong", Long.MAX_VALUE, memoryQueue.getMaxSize());
+        assertEquals("MaxSize wrong", Integer.MAX_VALUE, memoryQueue.getMaxSize());
     }
 
     public void testObserver() throws SinkException, SourceException {
@@ -395,17 +395,17 @@ public class TestTransactionalMemoryQueueWithDSTrx extends TestCase {
         manager.getTransactionManager().commit();
 
 
-        assertEquals("Wrong notify size for 1 put", Long.valueOf(6L), queueSize.get(0));
-        assertEquals("Wrong notify size for 2 put", Long.valueOf(10L), queueSize.get(1));
-        assertEquals("Wrong notify size for 1 take", Long.valueOf(9l), queueSize.get(2));
-        assertEquals("Wrong notify size for 2 take", Long.valueOf(4), queueSize.get(3));
-        assertEquals("Wrong notify size for 3 take", Long.valueOf(0), queueSize.get(4));
+        assertEquals("Wrong notify size for 1 put", Integer.valueOf(6), queueSize.get(0));
+        assertEquals("Wrong notify size for 2 put", Integer.valueOf(10), queueSize.get(1));
+        assertEquals("Wrong notify size for 1 take", Integer.valueOf(9), queueSize.get(2));
+        assertEquals("Wrong notify size for 2 take", Integer.valueOf(4), queueSize.get(3));
+        assertEquals("Wrong notify size for 3 take", Integer.valueOf(0), queueSize.get(4));
 
-        assertEquals("Wrong notify delta for 1 put", Long.valueOf(6L), delta.get(0));
-        assertEquals("Wrong notify delta for 2 put", Long.valueOf(4L), delta.get(1));
-        assertEquals("Wrong notify delta for 1 take", Long.valueOf(-1l), delta.get(2));
-        assertEquals("Wrong notify delta for 2 take", Long.valueOf(-5), delta.get(3));
-        assertEquals("Wrong notify delta for 3 take", Long.valueOf(-4), delta.get(4));
+        assertEquals("Wrong notify delta for 1 put", Integer.valueOf(6), delta.get(0));
+        assertEquals("Wrong notify delta for 2 put", Integer.valueOf(4), delta.get(1));
+        assertEquals("Wrong notify delta for 1 take", Integer.valueOf(-1), delta.get(2));
+        assertEquals("Wrong notify delta for 2 take", Integer.valueOf(-5), delta.get(3));
+        assertEquals("Wrong notify delta for 3 take", Integer.valueOf(-4), delta.get(4));
 
         queue.setObserver(null);
         queueSize.clear();
@@ -1948,7 +1948,7 @@ public class TestTransactionalMemoryQueueWithDSTrx extends TestCase {
     }
 
     public void test_Clone_Copy() throws SinkException, SourceException {
-        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.CLONE, 10l);
+        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.CLONE, 10);
 
         manager.getTransactionManager().begin();
         Event event = new Event("1");
@@ -1969,7 +1969,7 @@ public class TestTransactionalMemoryQueueWithDSTrx extends TestCase {
     }
 
     public void test_Non_Clone_Copy() throws SinkException, SourceException {
-        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.CLONE, 10l);
+        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.CLONE, 10);
 
         manager.getTransactionManager().begin();
         NonSerializableEvent event = new NonSerializableEvent("1");
@@ -1990,7 +1990,7 @@ public class TestTransactionalMemoryQueueWithDSTrx extends TestCase {
     }
 
     public void test_Non_Copy() throws SinkException, SourceException {
-        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.NONE, 10l);
+        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.NONE, 10);
 
         manager.getTransactionManager().begin();
         NonSerializableEvent event = new NonSerializableEvent("1");
@@ -2079,7 +2079,7 @@ public class TestTransactionalMemoryQueueWithDSTrx extends TestCase {
     }
 
     public void test_TakeCLONE_Rollback() throws SinkException, SourceException {
-        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.CLONE, 10l);
+        IQueue queue1 = new TransactionalMemoryQueue(manager, ObjectCloneType.CLONE, 10);
 
         manager.getTransactionManager().begin();
         Event event = new Event("1");
@@ -2219,6 +2219,6 @@ public class TestTransactionalMemoryQueueWithDSTrx extends TestCase {
     }
 
     protected TransactionalMemoryQueue createQueue() {
-        return new TransactionalMemoryQueue(manager, ObjectCloneType.SERIALIZE, 10l);
+        return new TransactionalMemoryQueue(manager, ObjectCloneType.SERIALIZE, 10);
     }
 }
